refactorizado de carpeta server
