--- server/views/index.html
@@ -1,14 +1,0 @@
-<html>
-    <head>
-      <title>Index Recursos</title>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1"> 
-    </head>
-    <!-- 3. Display the application -->
-    <body>
-      <h1>Index app</h1> 
-
-      <a href="http://localhost:8081/api/resources">Recursos API</a>
-    </body>
-  </html>
-
